+++ src/Views/Columns/Traits/Configuration/ColumnConfiguration.php
@@ -50,13 +50,6 @@ trait ColumnConfiguration
         return $this;
     }
 
-    public function setCustomSlug(string $customSlug): self
-    {
-        $this->customSlug = $customSlug;
-
-        return $this;
-    }
-
     public function setColumnLabelStatusDisabled(): self
     {
         $this->setColumnLabelStatus(false);
