+++ src/Views/Columns/Traits/Helpers/ComponentColumnHelpers.php
@@ -3,24 +3,4 @@
 namespace Rappasoft\LaravelLivewireTables\Views\Columns\Traits\Helpers;
 
 trait ComponentColumnHelpers
-{
-    public function getSlotCallback(): ?callable
-    {
-        return $this->slotCallback;
-    }
-
-    public function hasSlotCallback(): bool
-    {
-        return $this->slotCallback !== null;
-    }
-
-    public function getComponentView(): string
-    {
-        return $this->componentView;
-    }
-
-    public function hasComponentView(): bool
-    {
-        return isset($this->componentView);
-    }
-}
+{}
